requirements.txt
pytz
boto3
PyJWT
fastapi
uvicorn
pydantic
python-dotenv
